requirements.txt
fastapi[sqlalchemy]==0.115.0
sqlalchemy==2.0.35
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic[email-validator]>=2,<3

